Refactor to use the result summary step from the research framework.
